+++ FirebaseAI/Tests/Unit/Types/Imagen/ImageGenerationParametersTests.swift
@@ -34,7 +34,8 @@ final class ImageGenerationParametersTests: XCTestCase {
       personGeneration: nil,
       outputOptions: nil,
       addWatermark: nil,
-      includeResponsibleAIFilterReason: true
+      includeResponsibleAIFilterReason: true,
+      includeSafetyAttributes: true
     )
 
     let parameters = ImagenModel.imageGenerationParameters(
@@ -57,7 +58,8 @@ final class ImageGenerationParametersTests: XCTestCase {
       personGeneration: nil,
       outputOptions: nil,
       addWatermark: nil,
-      includeResponsibleAIFilterReason: true
+      includeResponsibleAIFilterReason: true,
+      includeSafetyAttributes: true
     )
 
     let parameters = ImagenModel.imageGenerationParameters(
@@ -95,7 +97,8 @@ final class ImageGenerationParametersTests: XCTestCase {
         compressionQuality: imageFormat.compressionQuality
       ),
       addWatermark: addWatermark,
-      includeResponsibleAIFilterReason: true
+      includeResponsibleAIFilterReason: true,
+      includeSafetyAttributes: true
     )
 
     let parameters = ImagenModel.imageGenerationParameters(
@@ -124,7 +127,8 @@ final class ImageGenerationParametersTests: XCTestCase {
       personGeneration: personFilterLevel.rawValue,
       outputOptions: nil,
       addWatermark: nil,
-      includeResponsibleAIFilterReason: true
+      includeResponsibleAIFilterReason: true,
+      includeSafetyAttributes: true
     )
 
     let parameters = ImagenModel.imageGenerationParameters(
@@ -170,7 +174,8 @@ final class ImageGenerationParametersTests: XCTestCase {
         compressionQuality: imageFormat.compressionQuality
       ),
       addWatermark: addWatermark,
-      includeResponsibleAIFilterReason: true
+      includeResponsibleAIFilterReason: true,
+      includeSafetyAttributes: true
     )
 
     let parameters = ImagenModel.imageGenerationParameters(
@@ -200,6 +205,7 @@ final class ImageGenerationParametersTests: XCTestCase {
     let outputOptions = ImageGenerationOutputOptions(mimeType: mimeType, compressionQuality: nil)
     let addWatermark = false
     let includeRAIReason = true
+    let includeSafetyAttributes = true
     let parameters = ImageGenerationParameters(
       sampleCount: sampleCount,
       storageURI: storageURI,
@@ -209,7 +215,8 @@ final class ImageGenerationParametersTests: XCTestCase {
       personGeneration: personGeneration,
       outputOptions: outputOptions,
       addWatermark: addWatermark,
-      includeResponsibleAIFilterReason: includeRAIReason
+      includeResponsibleAIFilterReason: includeRAIReason,
+      includeSafetyAttributes: includeSafetyAttributes
     )
 
     let jsonData = try encoder.encode(parameters)
@@ -220,6 +227,7 @@ final class ImageGenerationParametersTests: XCTestCase {
       "addWatermark" : \(addWatermark),
       "aspectRatio" : "\(aspectRatio)",
       "includeRaiReason" : \(includeRAIReason),
+      "includeSafetyAttributes" : \(includeSafetyAttributes),
       "negativePrompt" : "\(negativePrompt)",
       "outputOptions" : {
         "mimeType" : "\(mimeType)"
@@ -246,7 +254,8 @@ final class ImageGenerationParametersTests: XCTestCase {
       personGeneration: nil,
       outputOptions: nil,
       addWatermark: addWatermark,
-      includeResponsibleAIFilterReason: nil
+      includeResponsibleAIFilterReason: nil,
+      includeSafetyAttributes: nil
     )
 
     let jsonData = try encoder.encode(parameters)
@@ -272,7 +281,8 @@ final class ImageGenerationParametersTests: XCTestCase {
       personGeneration: nil,
       outputOptions: nil,
       addWatermark: nil,
-      includeResponsibleAIFilterReason: nil
+      includeResponsibleAIFilterReason: nil,
+      includeSafetyAttributes: nil
     )
 
     let jsonData = try encoder.encode(parameters)