+++ FirebaseAI/CHANGELOG.md
@@ -7,8 +7,12 @@
   tokens now, so your app will be ready to take advantage of replay
   protection when it becomes available for Firebase AI Logic.
 
+# 12.2.0
+- [feature] Added support for returning thought summaries, which are synthesized
+  versions of a model's internal reasoning process. (#15096)
+
 # 12.0.0
-- [added] Added support for Grounding with Google Search. (#15014)
+- [feature] Added support for Grounding with Google Search. (#15014)
 - [removed] Removed `CountTokensResponse.totalBillableCharacters` which was
   deprecated in 11.15.0. Use `totalTokens` instead. (#15056)