+++ Firestore/CHANGELOG.md
@@ -1,3 +1,6 @@
+# 12.1.0
+- [fixed] Fixed accidental removal of `pod "Firebase/Firestore"` for tvOS in 12.0.0.
+
 # 11.12.0
 - [fixed] Fixed the `null` value handling in `isNotEqualTo` and `notIn` filters.